ga_google/requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
deap==1.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-pubsub==0.39.1
googleapis-common-protos==1.5.6
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
idna==2.8
numpy==1.16.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
